+++ odoo/local-src/account_invoice_sent/models/account_invoice.py
@@ -2,7 +2,8 @@
 # Copyright 2015-2018 Jacques-Etienne Baudoux (BCIM) <je@bcim.be>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from odoo import models, fields
+from odoo import api, models, fields, _
+from odoo.addons.queue_job.job import job
 
 
 class AccountInvoice(models.Model):
@@ -16,3 +17,53 @@ class AccountInvoice(models.Model):
         readonly=True,
         related="partner_id.invoice_sending_method",
         string="Sending Method")
+
+    @api.multi
+    def _filter_send_invoice(self, sending_method=None):
+
+        def f_state(r):
+            return (
+                not r.sent and
+                r.state not in ('draft', 'proforma', 'proforma2')
+            )
+
+        def f_sending_method(r):
+            commercial = r.partner_id.commercial_partner_id
+            return commercial.invoice_sending_method == sending_method
+
+        def f_email(r):
+            commercial = r.partner_id.commercial_partner_id
+            return bool(commercial.email)
+
+        filters = [f_state]
+        if sending_method:
+            filters.append(f_sending_method)
+        if sending_method == 'email':
+            filters.append(f_email)
+
+        return self.filtered(
+            lambda r: all(f(r) for f in filters)
+        )
+
+    @job(default_channel='root.background.invoice')
+    def _generate_send_invoice(self, sending_method):
+        """Generate jobs to send invoices"""
+        invoices = self.exists()
+        invoices = invoices._filter_send_invoice(sending_method)
+        method_name = '_send_invoice_%s' % (sending_method,)
+        for invoice in invoices:
+            getattr(invoice.with_delay(), method_name)()
+
+    @job(default_channel='root.background.invoice')
+    def _send_invoice_email(self):
+        """Generate and send an invoice by email"""
+        # we need to apply the filter because the state may have
+        # changed since when we delayed the job
+        invoices = self.exists()._filter_send_invoice(sending_method='email')
+        if not invoices:
+            return
+        invoices.write({'sent': True})
+        template = self.env.ref('account.email_template_edi_invoice')
+        for invoice in invoices:
+            invoice.message_post(body=_("Invoice sent"))
+            template.send_mail(invoice.id)